permisos de los roles y sus vistas

+++ app/Http/Controllers/RoleController.php
@@ -5,18 +5,22 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class RoleController extends Controller
 {
+
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index(Request $request)
     {
+        $this->authorize('rol-list');
         $texto = $request->input('texto'); // variable del texto de busqueda
         $registros = Role::with('permissions')->where('name', 'like', "%{$texto}%")
             ->orderBy('id', 'desc')
-            ->paginate(2);
+            ->paginate(10);
 
         return view('role.index', compact('registros', 'texto'));
 
@@ -27,6 +31,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        $this->authorize('rol-create');
         $permissions = Permission::all();
 
         return view('role.action', compact('permissions'));
@@ -37,6 +42,7 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('rol-create');
         $request->validate([
             'name' => 'required|unique:roles,name',
             'permissions' => 'required|array',
@@ -60,6 +66,8 @@ class RoleController extends Controller
      */
     public function edit(string $id)
     {
+        $this->authorize('rol-edit');
+
         $registro = Role::findOrFail($id);
         $permissions = Permission::all();
 
@@ -71,13 +79,15 @@ class RoleController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $this->authorize('rol-edit');
         $registro = Role::findOrFail($id);
         $request->validate([
-            'name' => 'required|unique:roles,name'.$registro->id,
+            'name' => 'required|unique:roles,name,' . $registro->id,
             'permissions' => 'required|array',
         ]);
         $registro->update(['name' => $request->name]);
         $registro->syncPermissions($request->permissions);
+        return redirect()->route('roles.index')->with('mensaje', 'Registro '.$registro->name. 'actualizado correctamente');
     }
 
     /**
@@ -85,6 +95,7 @@ class RoleController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('rol-delete');
         $registro = Role::findOrFail($id);
         $registro->delete();
