+++ app/Http/Controllers/UserController.php
@@ -4,15 +4,18 @@ namespace App\Http\Controllers;
 use App\Http\Requests\UserRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 
 class UserController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index(Request $request)
     {
+        $this->authorize('user-list');
         $texto =$request->input('texto');
         $registros=User::where('name','like','%'.$texto.'%')
             ->orWhere('email','like','%'.$texto.'%')
@@ -27,6 +30,7 @@ class UserController extends Controller
      */
     public function create()
     {
+        $this->authorize('user-create');
         return view('usuario.action');
     }
 
@@ -35,6 +39,7 @@ class UserController extends Controller
      */
     public function store(UserRequest $request)
     {
+        $this->authorize('user-create');
         $registro=new User();
         $registro->name=$request->input('name');
         $registro->email=$request->input('email');
@@ -57,6 +62,7 @@ class UserController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('user-edit');
         $registro=User::findOrFail($id);
         return view('usuario.action', compact('registro'));
     }
@@ -66,6 +72,7 @@ class UserController extends Controller
      */
     public function update(UserRequest $request, $id)
     {
+        $this->authorize('user-edit');
         $registro=User::findOrFail($id);
         $registro->name=$request->input('name');
         $registro->email=$request->input('email');
@@ -80,12 +87,15 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('user-delete');
         $registro=User::findOrFail($id);
         $registro->delete();
         return redirect()->route('usuarios.index')->with('mensaje', $registro->name. ' eliminado satisfatoriamente.');
     }
 
-    public function toggleStatus(User $usuario){
+    public function toggleStatus(User $usuario)
+    {
+        $this->authorize('user-activate');
         $usuario->activo = !$usuario->activo;
         $usuario->save();
         return redirect()->route('usuarios.index')->with('mensaje', 'El estado del usuario fue actualizado satisfactoriamente');
